Match 2 functions in Menu/radio.c

+++ src/Menu/menu_radio_8004D2FC.c
@@ -7,6 +7,6 @@ extern menu_save_mode_data stru_8009F2D8;
 
 void menu_radio_8004D2FC(menu_save_mode_data *pSaveMode)
 {
-    init_radio_message_board_80040F74((int)&gMenuMan_800BD360);
+    init_radio_message_board_80040F74(&gMenuMan_800BD360);
     init_file_mode_8004D24C(pSaveMode, 0);
 }
